pyproject.toml
[tool.poetry]
name = "duda-discord-bot"
version = "0.1.0"
description = ""
authors = ["Matheus Moon"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.4"
discord = "2.3.2"
openai = "1.37.1"
python-decouple = "3.8"
asyncio = "3.4.3"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
